% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_visualization.R
\name{order_acc}
\alias{order_acc}
\title{Sort rows by behaviour label}
\usage{
order_acc(df_raw = NULL)
}
\arguments{
\item{df_raw}{A data.frame or tibble contains raw accelerometer data. In this
version, it should be arranged as x,y,z,x,y,z,...,label for each row and all
rows should have the same length. For dual-axial accelerometer data, it should
be arranged as x,y,x,y,...,label. And for single-axial, it's arranged as x,x,...,
label.}
}
\value{
A data.frame or tibble contains sorted raw accelerometer data.
}
\description{
Arrange the rows according to behaviour labels. Some
dataset might arrange same behaviours in different batches. For all other
functions in the rabc package data are expected to be presented in this ordered format.
}
\examples{
data(whitestork_acc)
whitestork_acc_muddled <- whitestork_acc[sample(1:nrow(whitestork_acc)),]
head(whitestork_acc_muddled[, ncol(whitestork_acc)], n = 15)
whitestork_acc_sorted <- order_acc(whitestork_acc_muddled)
head(whitestork_acc_sorted[, ncol(whitestork_acc)], n = 15)
}
