% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_visualization.R
\name{plot_acc}
\alias{plot_acc}
\title{Accelerometer data visualization}
\usage{
plot_acc(df_raw = NULL, axis_num = 3)
}
\arguments{
\item{df_raw}{A data.frame or tibble contains accelerometer data. In this
version, it should be arranged as x,y,z,x,y,z,...,label for each row and all
rows should have the same length. For dual-axial accelerometer data, it should
be arranged as x,y,x,y,...,label. And for single-axial, it's arranged as x,x,...,
label.}

\item{axis_num}{An integer scalar indicating number of axis in the dataset.
Default value is 3, which represents tri-axial accelerometer data. It only
accepts 1, 2 or 3.}
}
\description{
Use dygraph to plot all accelerameter data grouped by behaviour types. It provides
the user a convenient impression of how the ACC signal relates to different behaviours
and also could be used for data quality control.
}
\details{
It is highly recommended to input sorted accelerometer data by
\code{order_acc}. The x axis of this dygraph indicate the row sequence number
of sorted dataset. In this case, if any data is not consistent with its neighbours
which have the same behaviour labels, user can check observation record and decide
if this is caused by label mistake.
}
\examples{
data(whitestork_acc)
whitestork_acc_sorted <- order_acc(whitestork_acc)
plot_acc(df_raw = whitestork_acc_sorted)

plot_acc(df_raw = cbind(whitestork_acc_sorted[, seq(3,121,by = 3)],
                       whitestork_acc_sorted[,121,drop = FALSE]),
          axis_num = 1) #plot z axis data, column 121 contains behaviour labels
}
