% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Classification result check}
\usage{
plot_confusion_matrix(df_feature = NULL, vec_label = NULL)
}
\arguments{
\item{df_feature}{A data.frame or tibble contains the feature set for classification}

\item{vec_label}{A character vector that contains all behaviour type labels.}
}
\value{
A data.frame with 3 columns. Column 1 "obs": factor of observations.
Column 2 "pre": factor of predictions. Column 3 "id": sequence of observations same
as the input "vec_label".
}
\description{
Plot classification result confusion table and return classification result
data.frame.
}
\details{
In this function, input dataset will be randomly partitioned into 5
folds without replacement according to levels of behaviour labels (see \code{\link{createFolds}}
for more details). Each fold of the 5 folds will be used as a test set while the
rest data used as the training set. The classification model used in this function
is XGBoost. Parameters of the XGBoost model are set like the following:
n_rounds = 10, max_depth = 6, eta = 0.3, gamma = 0, colsample_bytree = 1,
min_child_weight = 1, subsample = 1. After each training and testing round of the
5 rounds, all original behaviour labels (observation) and predicted behaviours will be recorded.
In the confusion table plot, x axis "obs" denotes observations and y axis "pre"
denotes predictions. Dots are colored according to classification results. The
annotated numbers are number of cases in each observation and prediction group.
}
\examples{
pred <- plot_confusion_matrix(df_feature = df_time, vec_label = label_vec)
head(pred)
}
