% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_visualization.R
\name{plot_feature}
\alias{plot_feature}
\title{Features sequential visualization}
\usage{
plot_feature(df_feature = NULL, vec_label = NULL)
}
\arguments{
\item{df_feature}{A data.frame or tibble calculated by \code{calculate_feature_time}
or \code{calculate_feature_freq}.}

\item{vec_label}{A character vector that contains all behaviour labels.}
}
\description{
Use dygraph to plot sequential feature(s) grouped by behaviour types. This function
allows users to generate a visual impression of the feature values by behaviour types.
}
\details{
It is recommended to input labels from sorted acclerometer data by
\code{order_acc}.  The x axis of this dygraph indicate the row sequence number
of sorted dataset.
}
\examples{
data(whitestork_acc)
whitestork_acc_sorted <- order_acc(whitestork_acc)
df_time <- calculate_feature_time(whitestork_acc_sorted, winlen_dba = 11)
plot_feature(df_feature = df_time[, "ODBA", drop = FALSE],
              vec_label = whitestork_acc_sorted[,ncol(whitestork_acc_sorted)])
plot_feature(df_feature = df_time,
              vec_label = whitestork_acc_sorted[,ncol(whitestork_acc_sorted)])
}
