% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_visualization.R
\name{plot_feature_grouped}
\alias{plot_feature_grouped}
\title{Features group visualization}
\usage{
plot_feature_grouped(df_feature = NULL, vec_label = NULL, geom = "boxplot")
}
\arguments{
\item{df_feature}{A data.frame or tibble calculated by \code{calculate_feature_time} or
\code{calculate_feature_freq}.}

\item{vec_label}{A character vector that contains all behaviour type labels.}

\item{geom}{A character scalar to indicate type of plot. The default is "boxplot",
it also supports "density" and "violin".}
}
\description{
This function generates feature distribution grouped by behaviour types.
}
\details{
In this function, plots are arranged by \code{facet_wrap}. In each plot
window, plots are arranged into 3 rows multiply by 3 columns. If the input total
number of features is not 9 multiples, the plots in final plot window will
automatically zoom to fill the whole window. It will be necessary to use
"Previous plot" and "Next plot" buttons to navigate between plots if the input
number of features is greater than 9.
}
\examples{
data(whitestork_acc)
whitestork_acc_sorted <- order_acc(whitestork_acc)
df_time <- calculate_feature_time(whitestork_acc_sorted, winlen_dba = 11)
plot_feature_grouped(df_feature = df_time,
                   vec_label = whitestork_acc_sorted[,ncol(whitestork_acc_sorted)])

plot_feature_grouped(df_feature = df_time[, 1:15],
                   vec_label = whitestork_acc_sorted[,ncol(whitestork_acc_sorted)])
}
