% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{plot_selection_accuracy}
\alias{plot_selection_accuracy}
\title{Feature selection accuracy plot}
\usage{
plot_selection_accuracy(results = NULL)
}
\arguments{
\item{results}{A list. It should be the returned list of function \code{select_features}.}
}
\description{
Plot accuracies of selected features.
}
\details{
Red line and dots in the plot denotes classification accuracies of accumulated
 selected features. Grey bars indicate accuracy increase (labeled with numbers) of each
 selected feature.
}
\examples{
data(whitestork_acc)
whitestork_acc_sorted <- order_acc(whitestork_acc)
df_time <- calculate_feature_time(df_raw = whitestork_acc_sorted, winlen_dba = 11)
results <- select_features(df_feature = df_time[,1:6],
                         vec_label = whitestork_acc_sorted[,ncol(whitestork_acc_sorted)],
                         no_features = 3)
plot_selection_accuracy(results = results)
}
