% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{plot_wrong_classifications}
\alias{plot_wrong_classifications}
\title{Wrong classification visualization}
\usage{
plot_wrong_classifications(df_raw = NULL, axis_num = 3, df_result = NULL)
}
\arguments{
\item{df_raw}{A data.frame or tibble contains accelerometer data. In this
version, it should be arranged as x,y,z,x,y,z,...,label for each row and all
rows should have the same length. For dual-axial accelerometer data, it should
be arranged as x,y,x,y,...,label. And for single-axial, it's arranged as x,x,...,
label.}

\item{axis_num}{An integer scalar indicating number of axies in the dataset.}

\item{df_result}{A data.frame resulted from \code{result_check}}
}
\description{
Use dygraph to plot wrong classification bouts on all acceleration data.
}
\details{
The base dygraph is the same with plot by \code{ACC_visual}. All wrong
predictions in the data.frame from \code{result_check} are denoted with dotted
line annotated with the predicted behaviour label on top.
}
\examples{
pred <- plot_confusion_matrix(df = df_time, vec_label = label_vec)
plot_wrong_classifications(whitestork_acc_sorted, df_result = pred)
}
