% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{plot_UMAP}
\alias{plot_UMAP}
\title{UMAP visualization}
\usage{
plot_UMAP(df_time = NULL, df_freq = NULL, label_vec = NULL)
}
\description{
UMAP - Uniform Manifold Approximation and Projection. UMAP is a tool to embed
high dimensional data into low dimensions and preserves the global structure
of the data. In this package, UMAP is used to transform features (time or/and
frequency domain) into two dimensions.
}
\details{
There are three panels in this Shiny App, representing three functions.
Panel 1: "UMAP calculation and tuning" - evaluating whether ACC features represent
         behaviours.
Panel 2: "Feature visualization through UMAP" - how feature values vary across the
         two-dimensional UMAP plot.
Panel 3: "Custom features" - evaluating the performance of selected features in
         differentiation of ACC data.

Before calling \code{plot_UMAP}, one should make sure that variables "df_time" (
time domain feature set, a data.frame or tibble), "df_freq" (frequency domain
feature set, a data.frame or tibble) and "label_vec" (corresponding behaviours,
a character vector) exist.
}
\examples{
data(whitestork_acc)
whitestork_acc_sorted <- order_acc(whitestork_acc)
df_time <- calculate_feature_time(whitestork_acc_sorted, winlen_dba = 11)
df_freq <- calculate_feature_freq(whitestork_acc_sorted, samp_freq = 10.54)
labels <- whitestork_acc_sorted[, 121]
plot_UMAP(df_time, df_freq, labels)
}
